components/patina_performance: Add crate readme
Adds a readme to the crate. This makes it easier to understand the
component when looking at the  source code and keeps the
documentation closer to the code so its easier to maintain.

The page in the docs/ directory is removed to simplify finding
component info in the readme.

It also provides documentation in crates.io.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>