components/patina_debugger: Add crate readme
Adds a readme to the crate. This makes it easier to understand the
component when looking at the  source code and keeps the
documentation closer to the code so its easier to maintain.

It also provides documentation in crates.io.

The documentation that was in docs/src/dev/debugging.md is brought
here and supplemented with some additional information. That file
is referenced in several places so it is retained but points to
other debugger documentation.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>